* release/201204091700: (545 commits)
  Fixing a problem with M:M relationships and Propel's generated set* methods
  Making the names of the title classes uniform
  Adding the right sidebar banner on the homepage
  Make the video boxes scrollable
  Update MemberVideos widget
  [Implement] cartEmpty To do: http://goo.gl/5iVNN
  Column list classes fix
  Update _widgetCollectionCategories and _widgetMemberVideos
  Implement 040412_CQ_Collections to html
  Work on mocking the Marketplace and News pages
  Updated the test fixtures for migrations up to 1333895904
  If there are not SQL queries to run, Propel does not bump up the migration version
  Adding sidebar for the News module
  Adding migrations for the homepage editorial slots: homepage_carousel and homepage_showcase
  Removing wp-stats plugin and updating google-authenticator for WordPress
  Stop adding watermark to the thumbnails
  Fixing model classes for ContentCatgory and ContentCategoryArchive
  Move carousel classes to homepage.less
  Small fixes
  Updating neo4jphp to version 0.0.7
  ...

+++ apps/legacy/modules/manage/actions/actions.class.php
@@ -24,50 +24,77 @@ class manageActions extends cqActions
 
     $collector = $this->getCollector();
     $form = new CollectorEditForm($collector);
+    $shipping_rates_form = new ShippingRatesCollectionForm($collector, array(
+        'tainted_request_values' => $request->getParameter('shipping_rates_collection'),
+    ));
 
     if ($request->isMethod('post'))
     {
-      $form->bind($request->getParameter('collector'), $request->getFiles('collector'));
-      if ($form->isValid())
+      if ($request->hasParameter('collector'))
       {
-        $form->save();
-
-        // Clear the Geo Cache
-        CollectorGeocacheQuery::create()->filterByCollectorId($collector->getId())->delete();
-
-        $message = 'Your profile/account information was updated.';
-        if ($collectorEmail = $form->getOption('newEmail', false))
+        $form->bind($request->getParameter('collector'), $request->getFiles('collector'));
+        if ($form->isValid())
         {
-          $subject = $this->__('You have changed your email at CollectorsQuest.com');
-          $body = $this->getPartial(
-            'emails/collector_email_change',
-            array(
-              'collector'     => $collector,
-              'collectorEmail'=> $collectorEmail
-            )
-          );
+          $form->save();
 
-          if ($this->sendEmail($form->getValue('email'), $subject, $body))
+          // Clear the Geo Cache
+          CollectorGeocacheQuery::create()->filterByCollectorId($collector->getId())->delete();
 
-            $message .= ' Email verification sent to ' . $form->getValue('email');
+          $message = 'Your profile/account information was updated.';
+          if ($collectorEmail = $form->getOption('newEmail', false))
+          {
+            $subject = $this->__('You have changed your email at CollectorsQuest.com');
+            $body = $this->getPartial(
+              'emails/collector_email_change',
+              array(
+                'collector'     => $collector,
+                'collectorEmail'=> $collectorEmail
+              )
+            );
+
+            if ($this->sendEmail($form->getValue('email'), $subject, $body))
+
+              $message .= ' Email verification sent to ' . $form->getValue('email');
+          }
+
+          $this->getUser()->setFlash('success', $message);
+
+          // Send the profile data to Defensio to analyse
+          $collector->sendToDefensio('UPDATE');
+
+          $this->redirect('@manage_profile');
         }
-
-        $this->getUser()->setFlash('success', $message);
-
-        // Send the profile data to Defensio to analyse
-        $collector->sendToDefensio('UPDATE');
-
-        $this->redirect('@manage_profile');
+        else
+        {
+          $this->getUser()->setFlash('error', 'There were some problems, please take a look below.');
+        }
       }
-      else
+
+      if ($collector->getIsSeller() && $request->hasParameter('shipping_rates_collection'))
       {
-        $this->getUser()->setFlash('error', 'There were some problems, please take a look below.');
+        if ($shipping_rates_form->bindAndSave($request->getParameter('shipping_rates_collection')))
+        {
+          $this->getUser()->setFlash('success', 'Your shipping information was updated');
+          $this->redirect('@manage_profile');
+        }
+        else
+        {
+          $this->getUser()->setFlash('error', 'There were some problems, please take a look below.');
+        }
       }
     }
 
     // Make the Form and Collector available in the template
     $this->form = $form;
     $this->collector = $collector;
+    $this->collector_addresses = $collector->getCollectorAddresss();
+
+    if ($collector->getIsSeller())
+    {
+      // only make the shipping rates form available to the template if the
+      // collecto is a seller
+      $this->shipping_rates_form = $shipping_rates_form;
+    }
 
     $this->addBreadcrumb($this->__('Collectors'), '@collectors');
     $this->addBreadcrumb($this->__('Your Profile'));
@@ -122,7 +149,7 @@ class manageActions extends cqActions
     }
     else if (count($this->collections) == 0)
     {
-      $this->getUser()->setFlash('highlight', 'You have no collections yet! <a href="' . $this->generateUrl('collection_create') . '">Click here</a> to create one and start selling items.');
+      $this->getUser()->setFlash('highlight', 'You have no collections yet! <a href="' . $this->generateUrl('collection_create') . '">Click here</a> to create one!');
     }
 
     $this->addBreadcrumb($this->__('Collections'), '@collections');
@@ -190,8 +217,8 @@ class manageActions extends cqActions
 
           $this->getUser()->setFlash("success", $this->__('Changes were saved!'));
           $this->redirect($this->getController()->genUrl(array(
-              'sf_route' => 'manage_collection_by_slug',
-              'sf_subject' => $collection,
+            'sf_route'   => 'manage_collection_by_slug',
+            'sf_subject' => $collection,
           )));
         }
         catch (PropelException $e)
@@ -450,7 +477,10 @@ class manageActions extends cqActions
 
     $form = new ManageCollectiblesForm(
       $collectibles,
-      array('collector' => $this->getUser()->getCollector(), 'embedded_form_class' => 'CollectibleEditForm')
+      array(
+        'collector'           => $this->getUser()->getCollector(),
+        'embedded_form_class' => 'CollectibleEditForm'
+      )
     );
 
     if ($request->isMethod('post'))
@@ -463,6 +493,7 @@ class manageActions extends cqActions
       if ($form->isValid())
       {
         $collector = $this->getUser()->getCollector();
+
         try
         {
           foreach ($form->getValues() as $value)
@@ -472,12 +503,12 @@ class manageActions extends cqActions
             $collectible->setName($value['name']);
             $collectible->setDescription($value['description'], 'html');
             $collectible->setTags(is_array($value['tags']) ? implode(', ', $value['tags']) : $value['tags']);
-            $collectible->clearCollectibleForSales();
             // handle collectible-collection M:M relation
             $collectible->setCollections(CollectionQuery::create()
-              ->filterById($value['collection_collectible_list'], Criteria::IN)
-              ->find()
+                  ->filterById($value['collection_collectible_list'], Criteria::IN)
+                  ->find()
             );
+
             $collectible->save();
 
             if ($value['thumbnail'])
@@ -518,7 +549,8 @@ class manageActions extends cqActions
         }
         catch (PropelException $e)
         {
-          // currently just skip the errors
+          // currently just skip the errors //Errors should never be skipped
+          // throw $e;
         }
 
         $this->getUser()->setFlash('success', 'Collectible data saved');
@@ -585,4 +617,134 @@ class manageActions extends cqActions
 
     return $this->redirect('@manage_collections');
   }
+
+  public function executeShoppingOrders()
+  {
+    $q = ShoppingOrderQuery::create()
+       ->filterByCollector($this->getCollector());
+
+    $this->shopping_orders = $q->find();
+
+    return sfView::SUCCESS;
+  }
+
+  public function executeShoppingOrder()
+  {
+    /** @var $shopping_order ShoppingOrder */
+    $shopping_order = $this->getRoute()->getObject();
+    $this->forward404Unless($this->getCollector()->isOwnerOf($shopping_order));
+
+    return sfView::SUCCESS;
+  }
+
+  /**
+   * Action ResendEmailChange
+   *
+   * @param sfWebRequest $request
+   *
+   */
+  public function executeResendEmailChange(sfWebRequest $request)
+  {
+    /* @var $collectorEmail CollectorEmail */
+    $collectorEmail = $this->getCollector()->getLastEmailChangeRequest();
+    $this->forward404Unless($collectorEmail instanceof CollectorEmail);
+
+    $collector = $collectorEmail->getCollector();
+
+    $subject = $this->__('You have changed your email at CollectorsQuest.com');
+    $body = $this->getPartial(
+      'emails/collector_email_change',
+      array(
+        'collector'     => $collector,
+        'collectorEmail'=> $collectorEmail
+      )
+    );
+
+    if ($this->sendEmail($collectorEmail->getEmail(), $subject, $body))
+    {
+      $this->getUser()->setFlash('success', 'Email verification sent to ' . $collectorEmail->getEmail());
+    }
+    else
+    {
+      $this->getUser()->setFlash('error', 'Invalid email');
+    }
+
+    $this->redirect('@manage_profile');
+  }
+
+  public function executeAddNewAddress(sfWebRequest $request)
+  {
+    $address = new CollectorAddress();
+    $address->setCollector($this->getCollector());
+    $form = new FrontendCollectorAddressForm($address);
+
+    if (sfRequest::POST == $request->getMethod())
+    {
+      if ($form->bindAndSave($request->getParameter($form->getName())))
+      {
+        $this->redirect('@manage_profile#collector-marketplace');
+      }
+    }
+    $this->form = $form;
+
+    $this->addBreadcrumb($this->__('Collectors'), '@collectors');
+    $this->addBreadcrumb($this->__('Your Profile'), '@manage_profile');
+    $this->addBreadcrumb($this->__('Add a new address'));
+
+    $this->prependTitle($this->__('Add a New Address'));
+
+    return sfView::SUCCESS;
+  }
+
+  public function executeEditAddress(sfWebRequest $request)
+  {
+    $address = $this->getRoute()->getObject();
+    $this->forward404Unless($this->getUser()->isOwnerOf($address));
+
+    $form = new FrontendCollectorAddressForm($address);
+
+    if (sfRequest::POST == $request->getMethod())
+    {
+      if ($form->bindAndSave($request->getParameter($form->getName())))
+      {
+        $this->redirect('@manage_profile#collector-marketplace');
+      }
+    }
+
+    $this->form = $form;
+
+    $this->addBreadcrumb($this->__('Collectors'), '@collectors');
+    $this->addBreadcrumb($this->__('Your Profile'), '@manage_profile');
+    $this->addBreadcrumb($this->__('Edit address'));
+
+    $this->prependTitle($this->__('Edit Address'));
+
+    return sfView::SUCCESS;
+  }
+
+  public function executeDeleteAddress(sfWebRequest $request)
+  {
+    $address = $this->getRoute()->getObject();
+    $this->forward404Unless($this->getUser()->isOwnerOf($address));
+
+    if (sfRequest::DELETE == $request->getMethod())
+    {
+      $address->delete();
+      $this->getUser()->setFlash('success',
+        $this->__('You have successfully removed an address from your account.'));
+
+      return $this->redirect('@manage_profile#collector-marketplace');
+    }
+
+    $this->collector_address = $address;
+
+    $this->addBreadcrumb($this->__('Collectors'), '@collectors');
+    $this->addBreadcrumb($this->__('Your Profile'), '@manage_profile');
+    $this->addBreadcrumb($this->__('Confirm address deletion'));
+
+    $this->prependTitle($this->__('Confirm Address Deletion'));
+
+    return sfView::SUCCESS;
+  }
+
 }
